services/articles/requirements.txt
amqp==5.1.1
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
Django==4.2.2
django-filter==23.2
djangorestframework==3.14.0
idna==3.4
kombu==5.3.1
Markdown==3.4.3
pytz==2023.3
requests==2.31.0
sqlparse==0.4.4
urllib3==2.0.3
vine==5.0.0
